lab3/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.4.0
psycopg[binary]==3.2.6
PyYAML==6.0.2
